Resolved circular dependencies
